lint: clippy
refactor: Remove unused imports to comply with Clippy lint

This commit removes unused imports in `task7.rs` and `lib.rs` to adhere
to the Clippy lint rules. The imports `taskSafe`, `error::Error`,
`task_unsafe`, and `SemaphoreOption` are no longer needed and have been
safely eliminated. This refactoring improves code readability and
reduces clutter. No functional changes are introduced.

+++ src/signal.rs
@@ -59,7 +59,7 @@ pub fn _signal_action(sig: i32, flag: i32, action: extern "C" fn(i32)) -> Result
         sa_flags: flag,
     };
     unsafe { sigemptyset(&mut signal.sa_mask as *mut _) };
-    unsafe { sigaction(sig as i32, &signal as *const _, std::ptr::null_mut()) }.if_error()
+    unsafe { sigaction(sig, &signal as *const _, std::ptr::null_mut()) }.if_error()
 }
 
 pub fn signal_action(sig: Signal, flag: i32, action: extern "C" fn(i32)) -> Result<i32, Error> {
